PDF Generation code
The purpose of this code file is to enable functionality where, upon submitting a form created using the WordPress Contact Form 7 plugin (or any custom form), a PDF is generated and sent to the user via email. Additionally, the user can download the PDF as a letter. This solution can be used with any form plugin or custom form; you simply need to provide the field IDs and manage the button IDs accordingly.